+++ src/parse.c
@@ -12,7 +12,7 @@ static void	init_map(const t_info *info, const size_t line_count)
 	map->width = 0;
 	map->height = line_count;
 	map->height_map = (int **)malloc(sizeof(int *) * map->height);
-	map->color_map = (int **)malloc(sizeof(int *) * map->height);
+	map->color_map = (int64_t **)malloc(sizeof(int64_t *) * map->height);
 	if (map->height_map == NULL || map->color_map == NULL)
 		clear_before_exit(info, ERR_MSG_MALLOC, 0);
 }
